% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{assign_weights_lfc}
\alias{assign_weights_lfc}
\title{assign_weights_lfc}
\usage{
assign_weights_lfc(lfc.vec, quantiles, lfc.thr, weights = c(3, 1, -2))
}
\arguments{
\item{lfc.vec}{Vector of continuous values such as LFC. For discretization absolute values are conisdered.}

\item{quantiles}{Assigns weights based on quantiles, if set TRUE.}

\item{lfc.thr}{The vector of thresholds for discretization. Is treated as quantiles, if quantiles set TRUE, otherwise as absolute thresholds.}

\item{weights}{The vector of discrete weights for each possible case, Default: c(3, 1, -2)}
}
\value{
The vector of corresponding weights.
}
\description{
Assigns discrete weights based on continuous values.
}
