% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bootstrap}
\alias{bootstrap}
\title{bootstrap}
\usage{
bootstrap(
  ps,
  level.columns = NULL,
  rho.pos.thr = 0.8,
  rho.neg.thr = NULL,
  p.adj.thr = 0.01,
  n.bootstrap.batches = 50,
  corr.pair.presence = 1
)
}
\arguments{
\item{ps}{phyloseq object}

\item{level.columns}{column names based on which to separate into groups and calculate correlations}

\item{rho.pos.thr}{threshold which is used for filtering out pairs with lower positive correlation, Default: 0.8}

\item{rho.neg.thr}{threshold which is used for filtering out pairs with higher negative correlation, if not set -rho.pos.thr is taken, Default: NULL}

\item{p.adj.thr}{threshold which is used for filtering out pairs with larger p adjusted value, Default: 0.01}

\item{n.bootstrap.batches}{number of times to bootstrap, Default: 50}

\item{corr.pair.presence}{the ratio of bootstrap batches where the correlation pair is included, Default: 1}
}
\value{
dataframe of correlation pairs
}
\description{
Calculates pariwise correlations by bootstrapping
}
