% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_random_ps_subset}
\alias{generate_random_ps_subset}
\title{generate_random_ps_subset}
\usage{
generate_random_ps_subset(ps, level.columns = NULL)
}
\arguments{
\item{ps}{phyloseq object}

\item{level.columns}{column names based on which to subset}
}
\value{
subsetted phyloseq object
}
\description{
Generates a random subset of phyloseq object based on specified columns.
}
