% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_network}
\alias{plot_network}
\title{plot_network}
\usage{
plot_network(
  corr.table = NULL,
  network.object = NULL,
  taxa1.column.name = "taxa_X",
  taxa2.column.name = "taxa_Y",
  similarity.column.name = "rho"
)
}
\arguments{
\item{corr.table}{Dataframe of co-occurrence network, Default: NULL}

\item{network.object}{igraph object of the co-occurrence network, Default: NULL}

\item{taxa1.column.name}{The column name for the taxa 1 in the pairs, Default: 'taxa_X'}

\item{taxa2.column.name}{The column name for the taxa 2 in the pairs, Default: 'taxa_Y'}

\item{similarity.column.name}{The column name for similarity measure, Default: 'rho'}
}
\description{
Plots co-occurrence network.
}
