% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_subgraph}
\alias{plot_subgraph}
\title{plot_subgraph}
\usage{
plot_subgraph(network.subgraph.object, similarity.column.name = "rho")
}
\arguments{
\item{network.subgraph.object}{igraph object containing 'subgraph' attribute for vertices and edges.}

\item{similarity.column.name}{The column name for similarity measure, Default: 'rho'}
}
\description{
Plots subgraph network on co-occurrence network.
}
