% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rarefy}
\alias{rarefy}
\title{rarefy}
\usage{
rarefy(ps, sample.sizes.list, level.column)
}
\arguments{
\item{ps}{phyloseq object}

\item{sample.sizes.list}{list of sample sizes with names of groups in level.column}

\item{level.column}{column name of the sample data based on which to rarefy the phyloseq object, i.e. Study}
}
\value{
rarefied phyloseq object
}
\description{
Rarefies phyloseq object.
}
