% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run.batches}
\alias{run.batches}
\title{run.batches}
\usage{
run.batches(
  study,
  trt.column,
  cntrl.level,
  trt.level,
  experiment,
  ps.rarefied,
  network.object,
  tax.lvl,
  lfc.thr,
  lfc.weights,
  lfc.by.quantiles
)
}
\arguments{
\item{study}{study name}

\item{trt.column}{treatment column}

\item{cntrl.level}{control name}

\item{trt.level}{treatment name}

\item{experiment}{experiment name}

\item{ps.rarefied}{rarefied phyloseq object}

\item{network.object}{igraph network object of the co-occurrence network}

\item{tax.lvl}{Taxonomic level}

\item{lfc.thr}{The vector of thresholds for discretization. Is treated as quantiles, if quantiles set TRUE, otherwise as absolute thresholds.}

\item{lfc.weights}{The vector of discrete weights for each possible case.}

\item{lfc.by.quantiles}{Assigns weights based on quantiles, if set TRUE.}
}
\value{
named list.
}
\description{
run subgraph analysis for each batch
}
