% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{subgraph_analysis}
\alias{subgraph_analysis}
\title{subgraph_analysis}
\usage{
subgraph_analysis(
  ps,
  network.object,
  trt.column,
  cntrl.level,
  trt.level,
  lfc.by.quantiles,
  lfc.thr,
  lfc.weights,
  tax.lvl = "Species",
  subset.columns = list()
)
}
\arguments{
\item{ps}{phyloseq object}

\item{network.object}{igraph object of the co-occurrence network}

\item{trt.column}{The treatment column name.}

\item{cntrl.level}{The category name for control in the trt.column.}

\item{trt.level}{The category name for treatment in the trt.column.}

\item{lfc.by.quantiles}{Assigns weights based on quantiles, if set TRUE.}

\item{lfc.thr}{The vector of thresholds for discretization. Is treated as quantiles, if quantiles set TRUE, otherwise as absolute thresholds.}

\item{lfc.weights}{The vector of discrete weights for each possible case.}

\item{tax.lvl}{Taxonomic level, Default: 'Species'}

\item{subset.columns}{Named list for subsetting the phyloseq object based on sample data, where the names are column names and values are vectors of values to keep, Default: list()}
}
\value{
named list, containing igraph object of the network with subgraph attribute for vertices and edges, and ancombc results
}
\description{
Extract subgraph based on phyloseq object and co-occurrence network.
}
