% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{subgraph_solver_mwcsr}
\alias{subgraph_solver_mwcsr}
\title{subgraph_solver_mwcsr}
\usage{
subgraph_solver_mwcsr(graph)
}
\arguments{
\item{graph}{igraph object containing 'weights' attribute for vertices.}
}
\value{
the solved subgraph
}
\description{
Returns the solved subgraph.
}
